lab2 tests all passed finally

+++ src/kvsrv/common.go
@@ -1,10 +1,12 @@
 package kvsrv
 
 type PutAppendType = int
+
 const (
-	Puttype 	PutAppendType = iota
+	Puttype PutAppendType = iota
 	Appendtype
 )
+
 // Put or Append
 type PutAppendArgs struct {
 	Key   string
@@ -12,30 +14,30 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	PutOrAppend	PutAppendType
+	PutOrAppend PutAppendType
 
-	ClientId 		int
-	TransactionId 	int64
+	ClientId      int
+	TransactionId int
 }
 
 type PutAppendReply struct {
 	Value string
 
-	ClientId 		int
-	AckId 	int64
+	ClientId int
+	AckId    int
 }
 
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
 
-	ClientId 		int
-	TransactionId 	int64
+	ClientId      int
+	TransactionId int
 }
 
 type GetReply struct {
 	Value string
 
-	ClientId 		int
-	AckId 	int64
+	ClientId int
+	AckId    int
 }
